requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blosc2==3.6.1
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.45
gradio==5.38.1
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.15.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
ndindex==1.10.0
networkx==3.5
nose==1.3.7
numexpr==2.11.0
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.12.0.88
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
portalocker==3.2.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
semantic-version==2.10.0
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tables==3.10.2
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.2.18
timm==1.0.19
tokenizers==0.22.1
tomlkit==0.13.3
torchmetrics==1.8.0
tqdm==4.67.1
transformers==4.56.2
trimesh==4.7.1
triton==3.0.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yacs==0.1.8
yarl==1.20.1
--extra-index-url https://download.pytorch.org/whl/cu118
xformers==0.0.27.post2+cu118